% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\name{get_unhcr_palettes}
\alias{get_unhcr_palettes}
\title{Get names of all unique palettes provided in unhcr}
\usage{
get_unhcr_palettes(full = FALSE)
}
\arguments{
\item{full}{Whether to include full palette names (with suffixes, e.g. \code{_cont}) or just stubs}
}
\value{
Vector of palette name stubs or full names
}
\description{
Get names of all unique palettes provided in unhcr
}
\examples{
get_unhcr_palettes()
}
