% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_align}
\alias{left_align}
\title{Left align text for title in ggplot2 object}
\usage{
left_align(plot_name, pieces)
}
\arguments{
\item{plot_name}{}

\item{pieces}{}
}
\value{
ggplot2 object with aligned title
}
\description{
Left align text
}
\examples{
\dontrun{
ggpubr::ggarrange(left_align(plot1, c("subtitle", "title", "caption")), ncol = 1, nrow = 1)
}

}
\author{
https://github.com/bbc/bbplot/blob/master/R/finalise_plot.R
}
