% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_discrete_unhcr}
\alias{scale_colour_discrete_unhcr}
\alias{scale_fill_discrete_unhcr}
\alias{scale_color_opinionated_unhcr}
\alias{scale_colour_opinionated_unhcr}
\alias{scale_fill_opinionated_unhcr}
\alias{scale_color_diverging_unhcr}
\alias{scale_colour_diverging_unhcr}
\alias{scale_fill_diverging_unhcr}
\alias{scale_color_continuous_unhcr}
\alias{scale_colour_continuous_unhcr}
\alias{scale_fill_continuous_unhcr}
\title{Custom color and fill scales

Custom coloring and filling functions based on unique color palettes

Specific functions include:
\itemize{
\item{scale_(color/colour/fill)_discrete_unhcr}{Discrete palette with either fixed or dynamically extended number of shades}
\item{scale_(color/colour/fill)_opinionated_unhcr}{Discrete palette with specific values for "good", "bad", and "neutral"}
\item{scale_(color/colour/fill)_diverging_unhcr}{Continuous diverging color palette}
\item{scale_(color/colour/fill)_continuous_unhcr}{COntinuous color palette}
}}
\usage{
scale_color_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_colour_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_fill_discrete_unhcr(palette = "test", extend = FALSE, ...)

scale_color_opinionated_unhcr(palette = "test", ...)

scale_colour_opinionated_unhcr(palette = "test", extend = FALSE, ...)

scale_fill_opinionated_unhcr(palette = "test", ...)

scale_color_diverging_unhcr(palette = "test", ...)

scale_colour_diverging_unhcr(palette = "test", ...)

scale_fill_diverging_unhcr(palette = "test", ...)

scale_color_continuous_unhcr(palette = "test", ...)

scale_colour_continuous_unhcr(palette = "test", ...)

scale_fill_continuous_unhcr(palette = "test", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}
}
\description{
Custom color and fill scales

Custom coloring and filling functions based on unique color palettes

Specific functions include:
\itemize{
\item{scale_(color/colour/fill)_discrete_unhcr}{Discrete palette with either fixed or dynamically extended number of shades}
\item{scale_(color/colour/fill)_opinionated_unhcr}{Discrete palette with specific values for "good", "bad", and "neutral"}
\item{scale_(color/colour/fill)_diverging_unhcr}{Continuous diverging color palette}
\item{scale_(color/colour/fill)_continuous_unhcr}{COntinuous color palette}
}

Discrete palette with either fixed or dynamically extended number of shades

Discrete palette with specific values for "good", "bad", and "neutral"

Continuous diverging color palette

COntinuous color palette
}
\references{
https://en.wikipedia.org/wiki/Traffic_light
}
