% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\docType{data}
\name{test_pal_cont}
\alias{test_pal_cont}
\title{Define two colors for endpoints of continuous palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
test_pal_cont
}
\description{
Define two colors for endpoints of continuous palette
}
\keyword{datasets}
