% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_palette.R
\docType{data}
\name{test_pal_op}
\alias{test_pal_op}
\title{Define opinionated discrete palette (good, neutral, bad)}
\format{
An object of class \code{character} of length 3.
}
\usage{
test_pal_op
}
\description{
Define opinionated discrete palette (good, neutral, bad)
}
\keyword{datasets}
