% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_ppt.R
\name{unhcr_ppt}
\alias{unhcr_ppt}
\title{UNHCR ppt template}
\usage{
unhcr_ppt(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \code{html_document}}
}
\description{
Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
