% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_slide.R
\name{unhcr_slide}
\alias{unhcr_slide}
\title{UNHCR HTML slide}
\usage{
unhcr_slide(...)
}
\arguments{
\item{...}{additional arguments provided to \code{moon_reader}}
}
\value{
An R Markdown output format.
}
\description{
Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
