% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_bar.R
\name{unhcr_style_bar}
\alias{unhcr_style_bar}
\title{UNHCR ggplot2 theme for Bar chart}
\usage{
unhcr_style_bar()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for bar chart
}
\references{
https://github.com/gadenbuie/ggpomological/blob/master/R/theme_pomological.R
}
\seealso{
[ggplot2::theme]
}
\author{
Edouard Legoupil - with inspiration from bbc
}
