% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_bar_big.R
\name{unhcr_style_bar_big}
\alias{unhcr_style_bar_big}
\title{UNHCR ggplot2 theme for bar chart with big labels for powerpoints}
\usage{
unhcr_style_bar_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for bar chart
}
\examples{
{
library(ggplot2)
data <- data.frame(x = 1:10, y = 1:10)
ggplot(data, aes(x, y)) + geom_bar() +
unhcr_style_bar_big()
}


}
\references{
https://github.com/gadenbuie/ggpomological/blob/master/R/theme_pomological.R
}
\seealso{
[ggplot2::theme]
}
\author{
Edouard Legoupil - with inspiration from bbc
}
