% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_histo.R
\name{unhcr_style_histo}
\alias{unhcr_style_histo}
\title{UNHCR ggplot2 theme for histogram}
\usage{
unhcr_style_histo()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for histogram
}
\examples{
unhcr_style_histo()

}
\seealso{
[ggplot2::theme]
}
\author{
Edouard Legoupil - with inspiration from bbc
}
