% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_histo_big.R
\name{unhcr_style_histo_big}
\alias{unhcr_style_histo_big}
\title{UNHCR ggplot2 theme for histogramm  with big labels for powerpoints}
\usage{
unhcr_style_histo_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for histogram with big labels for powerpoints
}
\examples{
unhcr_style_histo_big()

}
\author{
Edouard Legoupil - with inspiration from bbc
}
