% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_map.R
\name{unhcr_style_map}
\alias{unhcr_style_map}
\title{UNHCR ggplot2 theme for maps}
\usage{
unhcr_style_map()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for maps
}
\examples{
\dontrun{
unhcr_style_map()
}

}
\author{
Edouard Legoupil -
}
