% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_map_big.R
\name{unhcr_style_map_big}
\alias{unhcr_style_map_big}
\title{UNHCR ggplot2 theme for maps with big labels for powerpoints}
\usage{
unhcr_style_map_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for maps with big labels for powerpoints
}
\examples{
\dontrun{
unhcr_style_map_big()
}

}
\author{
Edouard Legoupil -
}
