% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_scatter.R
\name{unhcr_style_scatter}
\alias{unhcr_style_scatter}
\title{UNHCR ggplot2 theme for scatter plot}
\usage{
unhcr_style_scatter()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for scatter plot
}
\examples{
\dontrun{
unhcr_style_scatter()
}

}
\author{
Edouard Legoupil - with inspiration from bbc
}
