% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_style_scatter_big.R
\name{unhcr_style_scatter_big}
\alias{unhcr_style_scatter_big}
\title{UNHCR ggplot2 theme for scatter plots  with big labels for powerpoints}
\usage{
unhcr_style_scatter_big()
}
\value{
Return UNHCR Style
}
\description{
Return ggplot2 styling for scatter plot with big labels for powerpoints
}
\examples{
\dontrun{
unhcr_style_scatter_big()
}

}
\seealso{
[ggplot2::theme]
}
\author{
Edouard Legoupil - with inspiration from bbc
}
