% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_word.R
\name{unhcr_word}
\alias{unhcr_word}
\title{UNHCR word template}
\usage{
unhcr_word(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \code{html_document}}
}
\value{
An R Markdown output format.
}
\description{
Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
