% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcrstyle.R
\docType{package}
\name{unhcRstyle}
\alias{unhcRstyle}
\title{\code{unhcRstyle} package}
\description{
A collection of template, palettes and themes for plots, documents, and slides
based on UNHCR branding
}
\details{
See the README on
\href{https://github.com/unhcr/unhcRstyle}{GitHub}
}
\references{
https://github.com/emilyriederer/Rtistic
}
